+++ src/compare/rtensor_rtensor_comparison.cc
@@ -18,17 +18,16 @@
 
 #include <algorithm>
 #include <tensor/tensor.h>
+#include <tensor/tensor/operators.h>
 
 namespace tensor {
 
-Booleans operator<=(const RTensor &a, const RTensor &b) {
-  tensor_assert(a.size() == b.size());
-  Booleans c(a.size());
-  Booleans::iterator ic = c.begin();
-  for (RTensor::const_iterator ia = a.begin(), ib = b.begin(); ia != a.end();
-       ia++, ib++, ic++)
-    *ic = (*ia) <= (*ib);
-  return c;
-}
+template Booleans operator<(const RTensor &a, const RTensor &b);
+template Booleans operator>(const RTensor &a, const RTensor &b);
+template Booleans operator==(const RTensor &a, const RTensor &b);
+template Booleans operator!=(const RTensor &a, const RTensor &b);
+template Booleans operator<=(const RTensor &a, const RTensor &b);
+template Booleans operator>=(const RTensor &a, const RTensor &b);
+template bool all_equal(const RTensor &a, const RTensor &b);
 
 }  // namespace tensor