+++ CHANGES.md
@@ -45,3 +45,5 @@ RTensor tensor = {{{1.0}, {2.0}}, {{3.0}, {4.0}}}
 * Similarly, `Indices` and `Booleans` no longer have a scalar constructor indicating the reserved size. Use the static function `Indices::empty` instead.
 
 * Member variables in class Sparse are now privated (the public declaration was a typo).
+
+* Comparison operators (`operator==`, `operator<=`, `all_equal`, etc) are now implemented using templates for ease of maintenance. This means that they no longer support passing `Vector` type objects or automatic conversion of arguments to `Tensor`.